+++ resources/views/includes/form/bool_select.blade.php
@@ -4,7 +4,7 @@
     @endforeach
 @endif
 
-<select class="filter form-control" name="{{ $field }}" value="{!! Request::only($field)[$field] !!}">
+<select class="filter form-control" name="{{ $field }}">
     <option value="">All</option>
     @if($object->count())
         @foreach(array_unique($selects) as $option)
@@ -22,8 +22,8 @@
             </option>
         @endforeach
     @else()
-        <option selected="selected" name="{!! Request::only($field)[$field] !!}">
-            {!! Request::only($field)[$field] !!}
+        <option selected="selected" name="{{ Request::only($field)[$field] }}">
+            {{ Request::only($field)[$field] }}
         </option>
     @endif
 </select>