+++ resources/views/includes/form/input_with_symbol.blade.php
@@ -5,7 +5,7 @@
             class="filter col-xs-12 pull-down"
             name="{{ $field  }}"
             type="text"
-            value="{!! Request::only($field)[$field] !!}"
+            value="{{ Request::only($field)[$field] }}"
         >
     </div>
 @endif