+++ CHANGELOG.md
@@ -12,6 +12,7 @@
 - Created links to view IP address management page inside of merchants view
 - Assign Location to Installation
 - Refactored RolesController
+- Partial Refund request, index and show.
 
 ## Bug Fixes
 - Removed unnecessary object checks in views